Enhance DomElement handling (see description)
Added possibility to pass single and arrays of DomElements to addEelements() and removeElements() functions.
These changes inevitably led to a change to the "selector" system. Instead of assigning a selector to each LaxElement as identification, each LaxElement now gets an unique ID. This unique ID is now also set as an attribute of the LaxElement's DomElement so that it can be removed by other "selector" arguments (previously the removeElements() function had to be passed exactly the same selector that was used in the addElements() function to remove the elements)